Inject a resource mapping into the client
In order to materialize resources with a working client (one that will
actually save) the client needs to have a mapping of properly
instantiated resources. This forces a stricter sequencing of
instantiation, but still lets each module be used in isolation. The
construction flow is largely top-down:

delighted -> client -> resources